api-cors-lambda-crud-dynamodb/requirements.txt
-e git+git@github.com:tomomano/aws-cdk-dojo.git@2cbc054ae059aaf8d7633cd2f13a417e4d6f9c38#egg=api_cors_lambda_crud_dynamodb&subdirectory=api-cors-lambda-crud-dynamodb
attrs==19.1.0
aws-cdk.assets==1.6.1
aws-cdk.aws-apigateway==1.6.1
aws-cdk.aws-applicationautoscaling==1.6.1
aws-cdk.aws-autoscaling-common==1.6.1
aws-cdk.aws-certificatemanager==1.6.1
aws-cdk.aws-cloudformation==1.6.1
aws-cdk.aws-cloudwatch==1.6.1
aws-cdk.aws-dynamodb==1.6.1
aws-cdk.aws-ec2==1.6.1
aws-cdk.aws-elasticloadbalancingv2==1.6.1
aws-cdk.aws-events==1.6.1
aws-cdk.aws-iam==1.6.1
aws-cdk.aws-kms==1.6.1
aws-cdk.aws-lambda==1.6.1
aws-cdk.aws-logs==1.6.1
aws-cdk.aws-route53==1.6.1
aws-cdk.aws-s3==1.6.1
aws-cdk.aws-s3-assets==1.6.1
aws-cdk.aws-sns==1.6.1
aws-cdk.aws-sqs==1.6.1
aws-cdk.aws-ssm==1.6.1
aws-cdk.core==1.6.1
aws-cdk.cx-api==1.6.1
aws-cdk.region-info==1.6.1
cattrs==0.9.0
jsii==0.15.2
publication==0.0.3
python-dateutil==2.8.0
six==1.12.0
typing-extensions==3.7.4
